+++ src/main/java/at/fhtw/swen3/persistence/entity/Parcel.java
@@ -1,136 +1,33 @@
 package at.fhtw.swen3.persistence.entity;
 
-import java.util.Objects;
+import at.fhtw.swen3.services.dto.RecipientDto;
 
-import at.fhtw.swen3.persistence.Recipient;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.v3.oas.annotations.media.Schema;
-import org.springframework.validation.annotation.Validated;
-import javax.validation.Valid;
-import javax.validation.constraints.*;
-
-/**
- * Parcel
- */
-@Validated
-@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2022-09-24T14:06:20.958Z[GMT]")
-
-
-public class Parcel   {
-  @JsonProperty("weight")
-  @Min(value = 1, message = "Weight must be greater than 0")
-  private Float weight = null;
-
-  @JsonProperty("recipient")
-  @NotNull
-  private Recipient recipient = null;
-
-  @JsonProperty("sender")
-  @NotNull
-  private Recipient sender = null;
-
-  public Parcel weight(Float weight) {
-    this.weight = weight;
-    return this;
-  }
-
-  /**
-   * Get weight
-   * @return weight
-   **/
-  @Schema(required = true, description = "")
-      @NotNull
+public class Parcel {
+    private Float weight = null;
+    private RecipientDto recipient = null;
+    private RecipientDto sender = null;
 
     public Float getWeight() {
-    return weight;
-  }
-
-  public void setWeight(Float weight) {
-    this.weight = weight;
-  }
-
-  public Parcel recipient(Recipient recipient) {
-    this.recipient = recipient;
-    return this;
-  }
-
-  /**
-   * Get recipient
-   * @return recipient
-   **/
-  @Schema(required = true, description = "")
-      @NotNull
-
-    @Valid
-    public Recipient getRecipient() {
-    return recipient;
-  }
-
-  public void setRecipient(Recipient recipient) {
-    this.recipient = recipient;
-  }
-
-  public Parcel sender(Recipient sender) {
-    this.sender = sender;
-    return this;
-  }
-
-  /**
-   * Get sender
-   * @return sender
-   **/
-  @Schema(required = true, description = "")
-      @NotNull
-
-    @Valid
-    public Recipient getSender() {
-    return sender;
-  }
-
-  public void setSender(Recipient sender) {
-    this.sender = sender;
-  }
-
-
-  @Override
-  public boolean equals(java.lang.Object o) {
-    if (this == o) {
-      return true;
+        return weight;
     }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
+
+    public void setWeight(Float weight) {
+        this.weight = weight;
     }
-    Parcel parcel = (Parcel) o;
-    return Objects.equals(this.weight, parcel.weight) &&
-        Objects.equals(this.recipient, parcel.recipient) &&
-        Objects.equals(this.sender, parcel.sender);
-  }
 
-  @Override
-  public int hashCode() {
-    return Objects.hash(weight, recipient, sender);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class Parcel {\n");
-    
-    sb.append("    weight: ").append(toIndentedString(weight)).append("\n");
-    sb.append("    recipient: ").append(toIndentedString(recipient)).append("\n");
-    sb.append("    sender: ").append(toIndentedString(sender)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(java.lang.Object o) {
-    if (o == null) {
-      return "null";
+    public RecipientDto getRecipient() {
+        return recipient;
     }
-    return o.toString().replace("\n", "\n    ");
-  }
+
+    public void setRecipient(RecipientDto recipient) {
+        this.recipient = recipient;
+    }
+
+    public RecipientDto getSender() {
+        return sender;
+    }
+
+    public void setSender(RecipientDto sender) {
+        this.sender = sender;
+    }
 }
