+++ src/main/java/at/fhtw/swen3/services/dto/ErrorDto.java
@@ -1,6 +1,79 @@
 package at.fhtw.swen3.services.dto;
 
-public class ErrorDto {
-    private String errorMessage = null;
+import java.util.Objects;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
+import org.springframework.validation.annotation.Validated;
 
+import javax.validation.constraints.*;
+
+/**
+ * Error
+ */
+@Validated
+@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2022-09-24T14:06:20.958Z[GMT]")
+
+
+public class ErrorDto {
+  @JsonProperty("errorMessage")
+  private String errorMessage = null;
+
+  public ErrorDto errorMessage(String errorMessage) {
+    this.errorMessage = errorMessage;
+    return this;
+  }
+
+  /**
+   * The error message.
+   * @return errorMessage
+   **/
+  @Schema(required = true, description = "The error message.")
+      @NotNull
+
+    public String getErrorMessage() {
+    return errorMessage;
+  }
+
+  public void setErrorMessage(String errorMessage) {
+    this.errorMessage = errorMessage;
+  }
+
+
+  @Override
+  public boolean equals(java.lang.Object o) {
+    if (this == o) {
+      return true;
+    }
+    if (o == null || getClass() != o.getClass()) {
+      return false;
+    }
+    ErrorDto error = (ErrorDto) o;
+    return Objects.equals(this.errorMessage, error.errorMessage);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(errorMessage);
+  }
+
+  @Override
+  public String toString() {
+    StringBuilder sb = new StringBuilder();
+    sb.append("class Error {\n");
+    
+    sb.append("    errorMessage: ").append(toIndentedString(errorMessage)).append("\n");
+    sb.append("}");
+    return sb.toString();
+  }
+
+  /**
+   * Convert the given object to string with each line indented by 4 spaces
+   * (except the first line).
+   */
+  private String toIndentedString(java.lang.Object o) {
+    if (o == null) {
+      return "null";
+    }
+    return o.toString().replace("\n", "\n    ");
+  }
 }
