+++ src/main/java/at/fhtw/swen3/services/dto/ParcelDto.java
@@ -1,33 +1,135 @@
 package at.fhtw.swen3.services.dto;
 
-import at.fhtw.swen3.persistence.Recipient;
+import java.util.Objects;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
+import org.springframework.validation.annotation.Validated;
+import javax.validation.Valid;
+import javax.validation.constraints.*;
+
+/**
+ * Parcel
+ */
+@Validated
+@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2022-09-24T14:06:20.958Z[GMT]")
+
 
 public class ParcelDto {
-    private Float weight = null;
-    private Recipient recipient = null;
-    private Recipient sender = null;
+  @JsonProperty("weight")
+  @Min(value = 1, message = "Weight must be greater than 0")
+  private Float weight = null;
+
+  @JsonProperty("recipient")
+  @NotNull
+  private RecipientDto recipient = null;
+
+  @JsonProperty("sender")
+  @NotNull
+  private RecipientDto sender = null;
+
+  public ParcelDto weight(Float weight) {
+    this.weight = weight;
+    return this;
+  }
+
+  /**
+   * Get weight
+   * @return weight
+   **/
+  @Schema(required = true, description = "")
+      @NotNull
 
     public Float getWeight() {
-        return weight;
-    }
+    return weight;
+  }
 
-    public void setWeight(Float weight) {
-        this.weight = weight;
-    }
+  public void setWeight(Float weight) {
+    this.weight = weight;
+  }
 
-    public Recipient getRecipient() {
-        return recipient;
-    }
+  public ParcelDto recipient(RecipientDto recipient) {
+    this.recipient = recipient;
+    return this;
+  }
 
-    public void setRecipient(Recipient recipient) {
-        this.recipient = recipient;
-    }
+  /**
+   * Get recipient
+   * @return recipient
+   **/
+  @Schema(required = true, description = "")
+      @NotNull
 
-    public Recipient getSender() {
-        return sender;
-    }
+    @Valid
+    public RecipientDto getRecipient() {
+    return recipient;
+  }
 
-    public void setSender(Recipient sender) {
-        this.sender = sender;
+  public void setRecipient(RecipientDto recipient) {
+    this.recipient = recipient;
+  }
+
+  public ParcelDto sender(RecipientDto sender) {
+    this.sender = sender;
+    return this;
+  }
+
+  /**
+   * Get sender
+   * @return sender
+   **/
+  @Schema(required = true, description = "")
+      @NotNull
+
+    @Valid
+    public RecipientDto getSender() {
+    return sender;
+  }
+
+  public void setSender(RecipientDto sender) {
+    this.sender = sender;
+  }
+
+
+  @Override
+  public boolean equals(java.lang.Object o) {
+    if (this == o) {
+      return true;
     }
+    if (o == null || getClass() != o.getClass()) {
+      return false;
+    }
+    ParcelDto parcel = (ParcelDto) o;
+    return Objects.equals(this.weight, parcel.weight) &&
+        Objects.equals(this.recipient, parcel.recipient) &&
+        Objects.equals(this.sender, parcel.sender);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(weight, recipient, sender);
+  }
+
+  @Override
+  public String toString() {
+    StringBuilder sb = new StringBuilder();
+    sb.append("class Parcel {\n");
+    
+    sb.append("    weight: ").append(toIndentedString(weight)).append("\n");
+    sb.append("    recipient: ").append(toIndentedString(recipient)).append("\n");
+    sb.append("    sender: ").append(toIndentedString(sender)).append("\n");
+    sb.append("}");
+    return sb.toString();
+  }
+
+  /**
+   * Convert the given object to string with each line indented by 4 spaces
+   * (except the first line).
+   */
+  private String toIndentedString(java.lang.Object o) {
+    if (o == null) {
+      return "null";
+    }
+    return o.toString().replace("\n", "\n    ");
+  }
 }
