+++ src/main/java/at/fhtw/swen3/services/mapper/ErrorMapper.java
@@ -1,11 +1,11 @@
 package at.fhtw.swen3.services.mapper;
 
-import at.fhtw.swen3.services.dto.ErrorDto;
+import at.fhtw.swen3.persistence.Error;
 import org.mapstruct.factory.Mappers;
 
 public interface ErrorMapper {
     ErrorMapper INSTANCE = Mappers.getMapper(ErrorMapper.class);
 
-    ErrorDto entityToDto(Error error);
-    Error DtoToEntity(ErrorDto error);
+    Error entityToDto(java.lang.Error error);
+    java.lang.Error DtoToEntity(Error error);
 }
