+++ src/main/java/at/fhtw/swen3/services/mapper/NewParcelInfoMapper.java
@@ -1,9 +1,8 @@
 package at.fhtw.swen3.services.mapper;
 
-import at.fhtw.swen3.persistence.entity.NewParcelInfo;
 import at.fhtw.swen3.services.dto.NewParcelInfoDto;
+import at.fhtw.swen3.persistence.entity.NewParcelInfo;
 import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
 import org.mapstruct.factory.Mappers;
 
 @Mapper
@@ -11,6 +10,6 @@ public interface NewParcelInfoMapper {
     NewParcelInfoMapper INSTANCE = Mappers.getMapper(NewParcelInfoMapper.class);
 
     //@Mapping(source="trackingId",target="trackingId") //falls namen unterschiedlich
-    NewParcelInfoDto entityToDto(NewParcelInfo newParcelInfo);
-    NewParcelInfo dtoToEntity(NewParcelInfoDto newParcelInfoDto);
+    NewParcelInfo entityToDto(NewParcelInfoDto newParcelInfo);
+    NewParcelInfoDto dtoToEntity(NewParcelInfo newParcelInfoDto);
 }
