+++ src/main/java/at/fhtw/swen3/services/mapper/ParcelMapper.java
@@ -1,7 +1,7 @@
 package at.fhtw.swen3.services.mapper;
 
-import at.fhtw.swen3.persistence.entity.Parcel;
 import at.fhtw.swen3.services.dto.ParcelDto;
+import at.fhtw.swen3.persistence.entity.Parcel;
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
 
@@ -9,6 +9,6 @@ import org.mapstruct.factory.Mappers;
 public interface ParcelMapper {
     ParcelMapper INSTANCE = Mappers.getMapper(ParcelMapper.class);
 
-    ParcelDto entityToDto(Parcel parcel);
-    Parcel dtoToEntity(ParcelDto parcelDto);
+    Parcel entityToDto(ParcelDto parcel);
+    ParcelDto dtoToEntity(Parcel parcelDto);
 }
