+++ src/main/java/at/fhtw/swen3/services/mapper/RecipientMapper.java
@@ -1,13 +1,13 @@
 package at.fhtw.swen3.services.mapper;
 
 
-import at.fhtw.swen3.persistence.Recipient;
 import at.fhtw.swen3.services.dto.RecipientDto;
+import at.fhtw.swen3.persistence.Recipient;
 import org.mapstruct.factory.Mappers;
 
 public interface RecipientMapper {
     RecipientMapper INSTANCE = Mappers.getMapper(RecipientMapper.class);
 
-    RecipientDto entityToDto(Recipient recipient);
-    Recipient DtoToEntity(RecipientDto recipient);
+    Recipient entityToDto(RecipientDto recipient);
+    RecipientDto DtoToEntity(Recipient recipient);
 }
