+++ src/main/java/at/fhtw/swen3/services/mapper/TrackingInformationMapper.java
@@ -1,7 +1,7 @@
 package at.fhtw.swen3.services.mapper;
 
-import at.fhtw.swen3.persistence.entity.TrackingInformation;
 import at.fhtw.swen3.services.dto.TrackingInformationDto;
+import at.fhtw.swen3.persistence.entity.TrackingInformation;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.factory.Mappers;
@@ -11,6 +11,6 @@ public interface TrackingInformationMapper {
     TrackingInformationMapper INSTANCE = Mappers.getMapper(TrackingInformationMapper.class);
 
     @Mapping(source="state",target="state") //falls namen unterschiedlich
-    TrackingInformationDto entityToDto(TrackingInformation trackingInformation);
-    TrackingInformation dtoToEntity(TrackingInformationDto trackingInformationDto);
+    TrackingInformation entityToDto(TrackingInformationDto trackingInformation);
+    TrackingInformationDto dtoToEntity(TrackingInformation trackingInformationDto);
 }
