+++ src/main/java/at/fhtw/swen3/services/mapper/TruckMapper.java
@@ -1,12 +1,12 @@
 package at.fhtw.swen3.services.mapper;
 
-import at.fhtw.swen3.persistence.Truck;
 import at.fhtw.swen3.services.dto.TruckDto;
+import at.fhtw.swen3.persistence.Truck;
 import org.mapstruct.factory.Mappers;
 
 public interface TruckMapper {
     TruckMapper INSTANCE = Mappers.getMapper(TruckMapper.class);
 
-    TruckDto entityToDto(Truck truck);
-    Truck DtoToEntity(TruckDto truck);
+    Truck entityToDto(TruckDto truck);
+    TruckDto DtoToEntity(Truck truck);
 }
