+++ src/main/java/at/fhtw/swen3/services/mapper/WarehouseMapper.java
@@ -1,12 +1,12 @@
 package at.fhtw.swen3.services.mapper;
 
-import at.fhtw.swen3.persistence.Transferwarehouse;
 import at.fhtw.swen3.services.dto.TransferwarehouseDto;
+import at.fhtw.swen3.persistence.Transferwarehouse;
 import org.mapstruct.factory.Mappers;
 
 public interface WarehouseMapper {
     TransferwarehouseMapper INSTANCE = Mappers.getMapper(TransferwarehouseMapper.class);
 
-    TransferwarehouseDto entityToDto(Transferwarehouse transferwarehouse);
-    Transferwarehouse DtoToEntity(TransferwarehouseDto Transferwarehouse);
+    Transferwarehouse entityToDto(TransferwarehouseDto transferwarehouse);
+    TransferwarehouseDto DtoToEntity(Transferwarehouse Transferwarehouse);
 }
