+++ src/test/java/at/fhtw/swen3/services/mapper/ParcelMapperTest.java
@@ -1,10 +1,8 @@
 package at.fhtw.swen3.services.mapper;
 
-import at.fhtw.swen3.persistence.Recipient;
-import at.fhtw.swen3.persistence.entity.NewParcelInfo;
-import at.fhtw.swen3.persistence.entity.Parcel;
-import at.fhtw.swen3.services.dto.NewParcelInfoDto;
+import at.fhtw.swen3.services.dto.RecipientDto;
 import at.fhtw.swen3.services.dto.ParcelDto;
+import at.fhtw.swen3.persistence.entity.Parcel;
 import org.junit.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
@@ -15,14 +13,14 @@ public class ParcelMapperTest {
 
     @Test
     public void entityToDto() {
-        Parcel parcel = new Parcel();
+        ParcelDto parcel = new ParcelDto();
         parcel.setWeight((float) 12.34);
-        parcel.setRecipient(new Recipient());
-        parcel.setSender(new Recipient());
+        parcel.setRecipient(new RecipientDto());
+        parcel.setSender(new RecipientDto());
         System.out.println(parcel);
 
         // AuthorMapper mapper = Mappers.getMapper(AuthorMapper.class);
-        ParcelDto parcelDto = ParcelMapper.INSTANCE.entityToDto(parcel);
+        Parcel parcelDto = ParcelMapper.INSTANCE.entityToDto(parcel);
 
         assertEquals(new Float(12.34), parcelDto.getWeight());
 
@@ -31,14 +29,14 @@ public class ParcelMapperTest {
 
     @Test
     public void dtoToEntity() {
-        ParcelDto parcelDto = new ParcelDto();
+        Parcel parcelDto = new Parcel();
         parcelDto.setWeight((float) 12.34);
-        parcelDto.setRecipient(new Recipient());
-        parcelDto.setSender(new Recipient());
+        parcelDto.setRecipient(new RecipientDto());
+        parcelDto.setSender(new RecipientDto());
         System.out.println(parcelDto);
 
         // AuthorMapper mapper = Mappers.getMapper(AuthorMapper.class);
-        Parcel parcel = ParcelMapper.INSTANCE.dtoToEntity(parcelDto);
+        ParcelDto parcel = ParcelMapper.INSTANCE.dtoToEntity(parcelDto);
 
         assertEquals(new Float(12.34), parcel.getWeight());
