+++ src/test/java/at/fhtw/swen3/services/mapper/TrackingInformationMapperTest.java
@@ -1,10 +1,7 @@
 package at.fhtw.swen3.services.mapper;
 
-import at.fhtw.swen3.persistence.Recipient;
-import at.fhtw.swen3.persistence.entity.Parcel;
-import at.fhtw.swen3.persistence.entity.TrackingInformation;
-import at.fhtw.swen3.services.dto.ParcelDto;
 import at.fhtw.swen3.services.dto.TrackingInformationDto;
+import at.fhtw.swen3.persistence.entity.TrackingInformation;
 import org.junit.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
@@ -15,26 +12,26 @@ public class TrackingInformationMapperTest {
 
     @Test
     public void entityToDto() {
-        TrackingInformation trackingInformation = new TrackingInformation();
-        trackingInformation.setState(TrackingInformation.StateEnum.INTRANSPORT);
+        TrackingInformationDto trackingInformation = new TrackingInformationDto();
+        trackingInformation.setState(TrackingInformationDto.StateEnum.INTRANSPORT);
         System.out.println(trackingInformation);
 
-        TrackingInformationDto trackingInformationDto = TrackingInformationMapper.INSTANCE.entityToDto(trackingInformation);
+        TrackingInformation trackingInformationDto = TrackingInformationMapper.INSTANCE.entityToDto(trackingInformation);
 
-        assertEquals(TrackingInformation.StateEnum.INTRANSPORT, trackingInformationDto.getState());
+        assertEquals(TrackingInformationDto.StateEnum.INTRANSPORT, trackingInformationDto.getState());
 
         System.out.println(trackingInformationDto);
     }
 
     @Test
     public void dtoToEntity() {
-        TrackingInformationDto trackingInformationDto = new TrackingInformationDto();
-        trackingInformationDto.setState(TrackingInformation.StateEnum.INTRANSPORT);
+        TrackingInformation trackingInformationDto = new TrackingInformation();
+        trackingInformationDto.setState(TrackingInformationDto.StateEnum.INTRANSPORT);
         System.out.println(trackingInformationDto);
 
-        TrackingInformation trackingInformation = TrackingInformationMapper.INSTANCE.dtoToEntity(trackingInformationDto);
+        TrackingInformationDto trackingInformation = TrackingInformationMapper.INSTANCE.dtoToEntity(trackingInformationDto);
 
-        assertEquals(TrackingInformation.StateEnum.INTRANSPORT, trackingInformation.getState());
+        assertEquals(TrackingInformationDto.StateEnum.INTRANSPORT, trackingInformation.getState());
 
         System.out.println(trackingInformation);
     }
